Redesign the IO API
Rewrote the API to allow writing any image of type T to any PNM
image. I.e., an image<float> could be written to a PBM, it is
up to the user if the conversion from float to [0,1] is desired.
Furthermore, any PNM image can be read to any image of type T.
I.e., a PPM image can be read to an image<float>, promoting the
integer values in the PPM image to float values.