Add implementation for reading PBM headers
Header reading currently is just a prototype. It is possible to
read a PBM header with arbitrary amounts of whitespace and
comments. The flexible PNM specification and the possibility to
contain headers makes parsing somewhat difficult. A context-free
grammar that describes the header may be possible.
Error handling for malformed headers is not complete.